# api/openapi.yml
openapi: 3.0.3
info:
  title: ym music service
  version: "0.1.0"
paths:
...
  /auth/refresh:
    post:
      security: []
      summary: update jwt tokens
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTokenRequest'
      responses:
        '200':
          description: token were successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /playlist:
    post:
      summary: creates new playlist
      operationId: createPlaylist
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PlaylistCreateRequest'
      responses:
        '201':
          description: playlist was created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistCreateResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: get all user's playlist
      operationId: getPlaylists
      responses :
        '200':
          description: playlists was fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlists'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /playlist/{playlistId}:
    post:
      summary: add track to playlist
      operationId: addTrackToPlaylist
      parameters:
        - in: path
          name: playlistId
          schema: 
            type: integer
            format: int32
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - track_id
              properties:
                track_id:
                  type: integer
                  format: int32
      responses:
        '200':
          description: track was successfully added to playlist
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: update playlist
      operationId: updatePlaylist
      parameters:
        - in: path
          name: playlistId
          schema:
            type: integer
            format: int32
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaylistResponse'
      responses:
        '200':
          description: playlist was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistWithTracksResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: get info about playlist
      operationId: getPlaylist
      parameters:
        - in: path
          name: playlistId
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: playlist data was successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistWithTracksResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: deletes playlist by id
      operationId: deletePlaylist
      parameters:
        - in: path
          name: playlistId
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: playlist was successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDeleteResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /playlist/{playlistId}/cover:
    post:
      description: uploads a new cover for the playlist
      operationId: uploadPlaylistCover
      parameters:
        - in: path
          name: playlistId
          schema:
            type: integer
            format: int32
          required: true
      requestBody:
        content:
          image/*:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: cover of the playlist was successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistCoverResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      description: deletes the playlist's cover, if there's none no-op
      operationId: deletePlaylistCover
      parameters:
        - in: path
          name: playlistId
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: cover of the playlist was successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistCoverResponse'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      description: fetching cover of the playlist
      operationId: getPlaylistCover
      parameters:
        - in: path
          name: playlistId
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: cover of the playlist was successfully fetched
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: no playlist was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required:
        - user_id
        - token_type
        - access_token
        - refresh_token
      properties:
        user_id:
          type: integer
          format: int32
        token_type:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    PlaylistCreateRequest:
      type: object
      required:
        - owner_id
        - playlist_name
      properties:
        owner_id:
          type: integer
          format: int32
        playlist_name:
          type: string
        playlist_cover:
          type: string
          format: binary
    PlaylistCreateResponse:
      type: object
      required:
        - playlist_id
      properties:
        playlist_id:
          type: integer
          format: int32
    PlaylistWithTracksResponse:
      type: object
      required:
        - playlist_id
        - playlist_name
        - tracks
      properties:
        playlist_id:
          type: integer
          format: int32
        playlist_name:
          type: string
        cover_url:
          type: string
        tracks:
          type: array
          items:
            type: integer
            format: int32
    PlaylistResponse:
      type: object
      required:
        - playlist_id
        - playlist_name
      properties:
        playlist_id:
          type: integer
          format: int32
        playlist_name:
          type: string
    Playlists:
      type: array
      items:
        $ref: '#/components/schemas/PlaylistResponse'
    PlaylistDeleteResponse:
      type: object
      required:
        - playlist_id
      properties:
        playlist_id:
          type: integer
          format: int32
    PlaylistCoverResponse:
      type: object
      required:
        - playlist_id
      properties:
        playlist_id:
          type: integer
          format: int32
...
security:
  - bearerAuth: []